broke up the god object
